conformance/buffimgs/cl_commandQueue_enqueueCopyImage.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueCopyImage conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<div style="position:absolute; left:0px; top:0px; visibility:hidden">
    <canvas id="dstImageCanvas">
</div>
<script>
try {
    debug("WebCLCommandQueue :: enqueueCopyImage() test case.\n");
    debug("( WebCLImage srcImage, WebCLImage dstImage, sequence <CLuint> srcOrigin, sequence <CLuint> dstOrigin, sequence <CLuint> region, optional sequence <WebCLEvent>? eventWaitList, optional WebCLEvent? event) \n");

    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLCommandQueue = wtu.createCommandQueue(webCLContext);
    var srcOrigin = [0, 0];
    var dstOrigin = [0, 0];
    var region = [10, 10];
    var imageWidth = 10;
    var imageHeight = 10;

    var webCLImageDescriptorReadWrite = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_WRITE, imageWidth, imageHeight);
    var srcImageReadWrite = webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite);
    var srcImageRead = webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorReadWrite);
    var srcImageWrite = webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite);

    // MEM_READ_WRITE destination image object.
    var dstImageReadWrite = webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageReadWrite, dstImageReadWrite, srcOrigin, dstOrigin, region);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageRead, dstImageReadWrite, srcOrigin, dstOrigin, region);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageWrite, dstImageReadWrite, srcOrigin, dstOrigin, region);");

    // MEM_WRITE_ONLY destination image object.
    var dstImageWrite = webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageReadWrite, dstImageWrite, srcOrigin, dstOrigin, region);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageRead, dstImageWrite, srcOrigin, dstOrigin, region);")
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageWrite, dstImageWrite, srcOrigin, dstOrigin, region);");

    // MEM_READ_ONLY destination image object.
    var dstImageRead = webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite);
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageReadWrite, dstImageRead, srcOrigin, dstOrigin, region);");
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageRead, dstImageRead, srcOrigin, dstOrigin, region);")
    shouldBeUndefined("webCLCommandQueue.enqueueCopyImage(srcImageWrite, dstImageRead, srcOrigin, dstOrigin, region);");

    // Testing for values of Image Offsets.
    var srcOriginInvalid = [imageWidth + 1, imageHeight + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(srcImageRead, dstImageWrite, srcOriginInvalid, dstOrigin, region);", "INVALID_VALUE");
    var dstOriginInvalid = [imageWidth + 1, imageHeight + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(srcImageRead, dstImageWrite, srcOrigin, dstOriginInvalid, region);", "INVALID_VALUE");
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(srcImageRead, dstImageWrite, srcOriginInvalid, dstOriginInvalid, region);", "INVALID_VALUE");

    // Testing for values of rectangular region being written.
    var regionInvalid = [imageWidth + 1, imageHeight + 1];
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(srcImageReadWrite, dstImageReadWrite, srcOrigin, dstOrigin, regionInvalid);", "INVALID_VALUE");

    // Testing for Exceptions.
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(invalid_object, dstImageReadWrite, srcOrigin, dstOrigin, region);", "INVALID_MEM_OBJECT");
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(srcImageReadWrite, invalid_object, srcOrigin, dstOrigin, region);", "INVALID_MEM_OBJECT");

    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(srcImageReadWrite, srcImageReadWrite, srcOrigin, srcOrigin, region);", "MEM_COPY_OVERLAP");

    var webCLContext2 = wtu.createContext();
    var webCLImageNew = webCLContext2.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite);
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(webCLImageNew, dstImageReadWrite, srcOrigin, dstOrigin, region);", "INVALID_CONTEXT");
    shouldThrowExceptionName("webCLCommandQueue.enqueueCopyImage(srcImageReadWrite, webCLImageNew, srcOrigin, dstOrigin, region);", "INVALID_CONTEXT");

    var webCLCommandQueueNew = wtu.createCommandQueue(webCLContext2);
    shouldThrowExceptionName("webCLCommandQueueNew.enqueueCopyImage(srcImageReadWrite, dstImageReadWrite, srcOrigin, dstOrigin, region);", "INVALID_CONTEXT");

    // FIXME :: Need to add test cases passing event parameters.

    // FIXME To add test cases to verify exception IMAGE_FORMAT_MISMATCH & INVALID_IMAGE_SIZE.

} catch (e) {
    testFailed("WebCLCommandQueue :: enqueueCopyImage() throws an exception : " + e.message);
}
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
